+++ swing/common-ui/src/test/java/is/codion/swing/common/ui/component/ComponentsTest.java
@@ -29,6 +29,7 @@ import is.codion.swing.common.ui.component.button.NullableCheckBox;
 import is.codion.swing.common.ui.component.button.ToggleButtonType;
 import is.codion.swing.common.ui.component.combobox.Completion;
 import is.codion.swing.common.ui.component.list.FilterList;
+import is.codion.swing.common.ui.component.multivalue.MultiValueInput;
 import is.codion.swing.common.ui.component.panel.InputPanelLayout;
 import is.codion.swing.common.ui.component.text.NumberField;
 import is.codion.swing.common.ui.component.text.TemporalField;
@@ -1045,10 +1046,10 @@ public final class ComponentsTest {
 	}
 
 	@Test
-	void listComboBox() {
+	void multiValueInput() {
 		Set<String> items = new HashSet<>(asList("one", "two", "three"));
-		ComponentValue<JComboBox<String>, Set<String>> componentValue =
-						Components.listComboBox()
+		ComponentValue<MultiValueInput<String>, Set<String>> componentValue =
+						Components.multiValueInput()
 										.component(Components.stringField().buildValue())
 										.value(items)
 										.buildValue();